Create switch for interference modality
