+++ src/automotive/model/TxTracker/channel_files/modified/multi-model-spectrum-channel.cc
